Dev branch fix.
So this is the dev branch.
Make sure we all pull request to this branch.

When a new version is ready we will create a new branch, supply the files to that branch from dev. And merge with Main. Then release version and signal auto updater a new version is available.